+++ packages/theme/src/lib.rs
@@ -1,8 +1,10 @@
 mod error;
 mod palette;
 mod theme;
+mod themes;
 
 pub use error::Error;
 pub use palette::Palette;
-pub use palette::{SUNO_DARK_PALETTE, SUNO_LIGHT_PALETTE};
 pub use theme::Theme;
+pub use themes::suno::SUNO_DARK_PALETTE;
+pub use themes::BUILTIN_PALETTES;