+++ packages/tui/src/widgets/input_password.rs
@@ -1,4 +1,5 @@
 use crate::widgets::input_field::InputField;
+use crate::widgets::spinner::render_spinner;
 use ratatui::{
     buffer::Buffer,
     layout::{Constraint, Direction, Layout, Position, Rect},
@@ -6,16 +7,17 @@ use ratatui::{
     widgets::{Block, Clear, Padding, Paragraph, Widget},
 };
 use std::sync::{Arc, RwLock};
-use suno_config::CONFIG;
+use suno_theme::Theme;
 
 #[derive(Debug)]
 pub struct InputPasswordWidget {
     pub state: Arc<RwLock<InputField>>,
+    pub theme: Theme,
 }
 
 impl Widget for &InputPasswordWidget {
     fn render(self, area: Rect, buf: &mut Buffer) {
-        let theme = CONFIG.theme();
+        let theme = self.theme;
         let mut state = self.state.write().unwrap();
 
         // Split area into two parts vertically for the main input field
@@ -101,8 +103,7 @@ impl Widget for &InputPasswordWidget {
 
         // Lock and show spinner when input is busy
         if state.is_busy() {
-            let spinner = state.spinner();
-            spinner.render(input_area[1], buf);
+            render_spinner(theme, state.spinner(), input_area[1], buf);
         }
 
         // Show invalid message when input is invalid